rustybuzz::hb

Module buffer

Modules

glyph_flag

Macros

foreach_cluster 🔒

foreach_grapheme 🔒

foreach_group 🔒

foreach_syllable 🔒

Structs

GlyphBuffer

A buffer that contains the results of the shaping process.

GlyphPosition

Holds the positions of the glyph in both horizontal and vertical directions.

GlyphPropsFlags

UnicodeBuffer

A buffer that contains an input string ready for shaping.

UnicodeProps

hb_buffer_t

hb_glyph_info_t

A glyph info.

Constants

CONTEXT_LENGTH 🔒

HB_BUFFER_CLUSTER_LEVEL_CHARACTERS

HB_BUFFER_CLUSTER_LEVEL_DEFAULT

HB_BUFFER_CLUSTER_LEVEL_MONOTONE_CHARACTERS

HB_BUFFER_CLUSTER_LEVEL_MONOTONE_GRAPHEMES

HB_BUFFER_SCRATCH_FLAG_DEFAULT

HB_BUFFER_SCRATCH_FLAG_HAS_BROKEN_SYLLABLE

HB_BUFFER_SCRATCH_FLAG_HAS_CGJ

HB_BUFFER_SCRATCH_FLAG_HAS_DEFAULT_IGNORABLES

HB_BUFFER_SCRATCH_FLAG_HAS_GLYPH_FLAGS

HB_BUFFER_SCRATCH_FLAG_HAS_GPOS_ATTACHMENT

HB_BUFFER_SCRATCH_FLAG_HAS_NON_ASCII

HB_BUFFER_SCRATCH_FLAG_HAS_SPACE_FALLBACK

HB_BUFFER_SCRATCH_FLAG_HAS_VARIATION_SELECTOR_FALLBACK

HB_BUFFER_SCRATCH_FLAG_SHAPER0

Functions

_cluster_group_func 🔒

Type Aliases

hb_buffer_cluster_level_t

hb_buffer_scratch_flags_t