rustybuzz::hb

Module ot_layout

OpenType layout.

Enums

TableIndex

Constants

IS_LIG_BASE 🔒

MAX_CONTEXT_LENGTH

MAX_NESTING_LEVEL

Traits

LayoutLookup

A lookup in a layout table.

LayoutTable

A lookup-based layout table (GSUB or GPOS).

LayoutTableExt

Functions

_hb_clear_substitution_flags

_hb_glyph_info_clear_default_ignorable 🔒

_hb_glyph_info_clear_ligated_and_multiplied 🔒

_hb_glyph_info_clear_substituted 🔒

_hb_glyph_info_get_general_category

_hb_glyph_info_get_lig_comp 🔒

_hb_glyph_info_get_lig_id 🔒

_hb_glyph_info_get_lig_num_comps 🔒

_hb_glyph_info_get_modified_combining_class

_hb_glyph_info_get_unicode_space_fallback_type 🔒

_hb_glyph_info_is_base_glyph 🔒

_hb_glyph_info_is_continuation 🔒

_hb_glyph_info_is_default_ignorable 🔒

_hb_glyph_info_is_hidden 🔒

_hb_glyph_info_is_ligature 🔒

_hb_glyph_info_is_mark 🔒

_hb_glyph_info_is_unicode_format 🔒

_hb_glyph_info_is_unicode_mark

_hb_glyph_info_is_unicode_space 🔒

_hb_glyph_info_is_variation_selector 🔒

_hb_glyph_info_is_zwj 🔒

_hb_glyph_info_is_zwnj 🔒

_hb_glyph_info_ligated 🔒

_hb_glyph_info_ligated_and_didnt_multiply 🔒

_hb_glyph_info_ligated_internal 🔒

_hb_glyph_info_multiplied 🔒

_hb_glyph_info_reset_continuation 🔒

_hb_glyph_info_set_continuation 🔒

_hb_glyph_info_set_general_category

_hb_glyph_info_set_lig_props_for_component 🔒

_hb_glyph_info_set_lig_props_for_ligature 🔒

_hb_glyph_info_set_lig_props_for_mark 🔒

_hb_glyph_info_set_modified_combining_class 🔒

_hb_glyph_info_set_unicode_space_fallback_type 🔒

_hb_glyph_info_set_variation_selector 🔒

_hb_glyph_info_substituted 🔒

_hb_grapheme_group_func 🔒

_hb_ot_layout_reverse_graphemes

_hb_ot_layout_set_glyph_props

apply_backward 🔒

apply_forward 🔒

apply_layout_table

Applies the lookups in the given GSUB or GPOS table.

apply_string 🔒

hb_ot_layout_has_cross_kerning

hb_ot_layout_has_glyph_classes

hb_ot_layout_has_kerning

hb_ot_layout_has_machine_kerning

hb_ot_layout_substitute_start

Called before substitution lookups are performed, to ensure that glyph class and other properties are set on the glyphs in the buffer.