Struct rustybuzz::GlyphInfo

#[repr(C)]
pub struct GlyphInfo {
    pub glyph_id: u32,
    pub cluster: u32,
    /* private fields */
}

A glyph info.

Fields
glyph_id: u32

A selected glyph.

Guarantee to be <= u16::MAX.

cluster: u32

An index to the start of the grapheme cluster in the original string.

Read more on clusters.

Implementations

impl hb_glyph_info_t

pub fn unsafe_to_break(&self) -> bool

Indicates that if input text is broken at the beginning of the cluster this glyph is part of, then both sides need to be re-shaped, as the result might be different.

On the flip side, it means that when this flag is not present, then itโs safe to break the glyph-run at the beginning of this cluster, and the two sides represent the exact same result one would get if breaking input text at the beginning of this cluster and shaping the two sides separately. This can be used to optimize paragraph layout, by avoiding re-shaping of each line after line-breaking, or limiting the reshaping to a small piece around the breaking point only.