Crate rustybuzz

A complete harfbuzz shaping algorithm port to Rust.

Re-exports

pub use ttf_parser;

Modules

script

Predefined scripts.

Structs

Face

A font face handle.

Feature

A feature tag with an accompanying range specifying on which subslice of shapes input it should be applied.

GlyphBuffer

A buffer that contains the results of the shaping process.

GlyphInfo

A glyph info.

GlyphPosition

Holds the positions of the glyph in both horizontal and vertical directions.

Language

A script language.

Script

A text script.

SerializeFlags

Flags used for serialization with a BufferSerializer.

Tag

A 4-byte tag.

UnicodeBuffer

A buffer that contains an input string ready for shaping.

Variation

A font variation.

Enums

BufferClusterLevel

A cluster level.

Direction

Defines the direction in which text is to be read.

Functions

shape

Shapes the buffer content using provided font and features.