Function rustybuzz::shape_with_plan

pub fn shape_with_plan(
    face: &Face<'_>,
    plan: &ShapePlan,
    buffer: UnicodeBuffer
) -> GlyphBuffer

Shapes the buffer content using the provided font and plan.

Consumes the buffer. You can then run GlyphBuffer::clear to get the UnicodeBuffer back without allocating a new one.

It is up to the caller to ensure that the shape plan matches the properties of the provided buffer, otherwise the shaping result will likely be incorrect.

Panics

Will panic when debugging assertions are enabled if the buffer and plan have mismatched properties.